src/requirement.txt
Khayyam==3.0.17
PyYAML==6.0.1
